requirements.txt
numpy>=1.26
pydantic>=2.7
pydantic-settings>=2.2
PyYAML>=6.0.1
opencv-python>=4.9
Pillow>=10.3
SQLAlchemy>=2.0
alembic>=1.13
rich>=13.7
typer>=0.12
python-dotenv>=1.0
pytest>=8.2
pytest-cov>=5.0
ruff>=0.5
mypy>=1.10
types-PyYAML
ipykernel>=6.29
matplotlib>=3.8
umap-learn>=0.5.11
python-docx>=1.1.0
